% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrmesh-package.R, R/agrmesh.R
\docType{package}
\name{agrmesh-package}
\alias{agrmesh-package}
\title{agrmesh: R Interface for AMGSDS}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R interface to access the Agro-Meteorological Grid Square Data system operated by NARO. See \url{https://amu.rd.naro.go.jp/} for more details of the system.

agrmesh provides R CUI and GUI tools for spatial meteorolgoical data analyses
}
\author{
\strong{Maintainer}: Keach Murakami \email{keach.murakami@gmail.com} (\href{https://orcid.org/0000-0001-8150-9535}{ORCID})

Other contributors:
\itemize{
  \item Yuichi Nagasaki [contributor]
}

}
\keyword{internal}
