% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{amgsds_config}
\alias{amgsds_config}
\title{Verify and initialize id and pw for AMGSDS system}
\usage{
amgsds_config(mode = .pkg_config$active_mode, .force = FALSE)
}
\arguments{
\item{mode}{chr. AMGSDS system mode, "oracle" or "classic" (for commercial users)}

\item{.force}{logical. pass verification for offline .onLoad}
}
\description{
Verify and initialize id and pw for AMGSDS system
}
