% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{area2code}
\alias{area2code}
\title{(lat range, lon range) -> covered mesh codes}
\usage{
area2code(lats, lons)
}
\arguments{
\item{lats}{latitudes}

\item{lons}{longitudes}
}
\description{
Convert a lat/lon bbox to Japanese mesh codes that covered by the bbox
}
