% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_latlon_range}
\alias{check_latlon_range}
\title{Confirm lat/lon ranges are in the scope}
\usage{
check_latlon_range(lats, lons)
}
\arguments{
\item{lats}{latitudes}

\item{lons}{longitudes}
}
\value{
stop if lats or lons protrude from the supported area
}
\description{
Confirm lat/lon ranges are in the scope
}
