% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_scenario_availability}
\alias{check_scenario_availability}
\title{Confirm scenario availability}
\usage{
check_scenario_availability(.times, .element, .model, .RCP)
}
\arguments{
\item{.times}{time range}

\item{.element}{element}

\item{.model}{model}

\item{.RCP}{RCP}
}
\value{
stop if either elements, model, or RCP is not supported by the scenario
}
\description{
Confirm scenario availability
}
