% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_source_availability}
\alias{check_source_availability}
\title{Confirm data availability}
\usage{
check_source_availability(.source, .element, .times, .clim)
}
\arguments{
\item{.source}{source}

\item{.element}{element}

\item{.times}{time range}

\item{.clim}{is climatological}
}
\value{
stop if either elements or times are not supported by the source data
}
\description{
Confirm data availability
}
