% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{check_variables}
\alias{check_variables}
\title{Extract a single variable column name}
\usage{
check_variables(tbl, element_index = 1)
}
\arguments{
\item{tbl}{AMGSDS tibble}

\item{element_index}{index of element to be shown}
}
\description{
Extract a single variable column name
}
