% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code2point}
\alias{code2point}
\title{mesh code -> (lat, lon)}
\usage{
code2point(code)
}
\arguments{
\item{code}{mesh code(s)}
}
\value{
lat/lon of mesh center(s)
}
\description{
Convert Japanese mesh code to lat/lon of the center of the mesh
}
