% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{download_netcdf}
\alias{download_netcdf}
\title{Download NetCDF files from AMGSDS server}
\usage{
download_netcdf(
  amgsds_path,
  outdir,
  server = "amd.rd.naro.go.jp/opendap",
  .silent = TRUE
)
}
\arguments{
\item{amgsds_path}{path object generated by generate_path function}

\item{outdir}{local directory to store downloaded files}

\item{server}{URL for OPeNDAP server (experimental)}

\item{.silent}{suppress download messages if \code{TRUE}}
}
\description{
Download NetCDF files from AMGSDS server
}
