% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fold_tibble}
\alias{fold_tibble}
\title{Conver long-format tibble to array}
\usage{
fold_tibble(amgsds_tibble)
}
\arguments{
\item{amgsds_tibble}{AMGSDS tibble object (output = "tibble")}
}
\description{
Conver long-format tibble to array
}
