% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{generate_path}
\alias{generate_path}
\title{Generate paths to AMGSDS data.}
\usage{
generate_path(
  times,
  lats,
  lons,
  element,
  source = "daily",
  model = NULL,
  RCP = NULL,
  is_clim = FALSE
)
}
\arguments{
\item{times}{A time range to extract.}

\item{lats}{Latitude range to extract.}

\item{lons}{longitude range to extract}

\item{element}{name of element to extract}

\item{source}{data source}

\item{model}{climate simulation model for scenario data}

\item{RCP}{Representative Concentration Pathways for scenario data}

\item{is_clim}{return climatological normals if \code{TRUE}}
}
\description{
Generate relative paths to NetCDF files in AMGSDS Server and local directory.
}
