% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_identity_cloud.R
\name{get_access_token}
\alias{get_access_token}
\title{Simplified token retrieval with automatic refresh}
\usage{
get_access_token()
}
\description{
Simplified token retrieval with automatic refresh
}
