% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{load_amgsds}
\alias{load_amgsds}
\title{Load data from local directory}
\usage{
load_amgsds(
  times = Sys.time(),
  lats,
  lons,
  elements,
  mode = "point",
  source = "daily",
  output = "tibble",
  model = "MIROC5",
  RCP = "RCP8.5",
  is_clim = FALSE,
  .silent = TRUE,
  localdir,
  autodownload = FALSE
)
}
\arguments{
\item{times}{time range to extract}

\item{lats}{latitude range to extract}

\item{lons}{longitude range to extract}

\item{elements}{names of element to extract}

\item{mode}{switch point/area/whole-mesh extraction}

\item{source}{data source}

\item{output}{format of output object}

\item{model}{climate simulation model for scenario data}

\item{RCP}{Representative Concentration Pathways for scenario data}

\item{is_clim}{logical. return climatological normals if \code{TRUE}}

\item{.silent}{logical. suppress download message if \code{TRUE}}

\item{localdir}{localdir directory for downloaded NetCDF files}

\item{autodownload}{logical. download lacking NetCDF files if \code{TRUE}}
}
\description{
Load data from local directory
}
