% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_identity_cloud.R
\name{load_tokens}
\alias{load_tokens}
\title{Load oracle tokens}
\usage{
load_tokens(OCI_TOKEN_FILE = get_config_value("OCI_TOKEN_FILE"))
}
\arguments{
\item{OCI_TOKEN_FILE}{path to the Oracle token file}
}
\description{
Load oracle tokens
}
