% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonempty_meshes}
\alias{nonempty_meshes}
\title{List of primary mesh codes with valid grid values}
\format{
A numeric vector with 304 meshes:
}
\usage{
nonempty_meshes
}
\description{
List of primary mesh codes with valid grid values
}
\keyword{datasets}
