% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonempty_meshes_fewer}
\alias{nonempty_meshes_fewer}
\title{List of primary mesh codes with valid grid values in scenario datasets}
\format{
A numeric vector with 166 meshes:
}
\usage{
nonempty_meshes_fewer
}
\description{
List of primary mesh codes with valid grid values in scenario datasets
}
\keyword{datasets}
