% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_identity_cloud.R
\name{pll_for_tokens}
\alias{pll_for_tokens}
\alias{poll_for_tokens}
\title{authorize device via Oracle Multi-Factor authentication}
\usage{
poll_for_tokens(device_code, interval, expires_in)
}
\arguments{
\item{expires_in}{}
}
\description{
authorize device via Oracle Multi-Factor authentication
}
