% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{plot2d_leaflet}
\alias{plot2d_leaflet}
\title{Plotting AMGSDS variables on interactive leaflet map}
\usage{
plot2d_leaflet(
  amgsds_tibble,
  element_index = 1,
  time_index = 1,
  alpha = 0.7,
  pallete = "viridis",
  basemap = leaflet:::providers$Wikimedia,
  thin = 0.01
)
}
\arguments{
\item{amgsds_tibble}{AMGSDS tibble object}

\item{element_index}{index of element to be shown}

\item{time_index}{index of time to be shown}

\item{alpha}{transparency of tiles (0 = fully transparent, 1 = opaque)}

\item{pallete}{color palette}

\item{basemap}{source of background map}

\item{thin}{thinning factor to avoid hanging (1 = show all meshes, 0.05 = show 5 \% of available meshes)}
}
\description{
Plotting AMGSDS variables on interactive leaflet map
}
