% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{plot2d_shape}
\alias{plot2d_shape}
\title{Plotting AMGSDS variables with shape map}
\usage{
plot2d_shape(
  amgsds_tibble,
  element_index = 1,
  time_index = 1,
  alpha = 0.5,
  pref_code
)
}
\arguments{
\item{amgsds_tibble}{AMGSDS tibble object}

\item{element_index}{index of element to be shown}

\item{time_index}{index of time to be shown}

\item{alpha}{transparency of tiles (0 = fully transparent, 1 = opaque)}

\item{pref_code}{pref id, refer to jpndistrict::jpnprefs}
}
\description{
Plotting AMGSDS variables with shape map
}
