% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{point2code}
\alias{point2code}
\title{(lat, lon) -> covering mesh code}
\usage{
point2code(lats, lons, simple = TRUE)
}
\arguments{
\item{lats}{latitudes}

\item{lons}{longitudes}

\item{simple}{(internal use)}
}
\value{
mesh code
}
\description{
Convert lat/lon to Japanese mesh code that covers the site
}
