% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2d.R
\name{polygonize}
\alias{polygonize}
\title{Create polygon shape geometry from data with latitude and longitude}
\usage{
polygonize(amgsds_tibble)
}
\arguments{
\item{amgsds_tibble}{AMGSDS object}
}
\description{
Create polygon shape geometry from data with latitude and longitude
}
