% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preview_dataset}
\alias{preview_dataset}
\title{Preview elements and time ranges covered by AMGSDS source dataset}
\usage{
preview_dataset(source = "daily", internal_use = FALSE)
}
\arguments{
\item{source}{type of source: daily, hourly, geo, or scenario}

\item{internal_use}{set TRUE for internal data check functions}
}
\description{
Preview elements and time ranges covered by AMGSDS source dataset
}
