% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{read_netcdf}
\alias{read_netcdf}
\title{Read NetCDF files from local directory}
\usage{
read_netcdf(amgsds_path, output = "tibble", localdir)
}
\arguments{
\item{amgsds_path}{path object generated by generate_path function}

\item{output}{format of output object}

\item{localdir}{local directory for downloaded NetCDF files}
}
\description{
Read NetCDF files from local directory
}
