% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_identity_cloud.R
\name{refresh_access_token}
\alias{refresh_access_token}
\title{Refresh access token using refresh token}
\usage{
refresh_access_token(tokens)
}
\description{
Refresh access token using refresh token
}
