% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{replace_stop_words}
\alias{replace_stop_words}
\title{Replace stop words in amgsds paths}
\usage{
replace_stop_words(str)
}
\arguments{
\item{str}{One or more character vectors.}
}
\description{
Replace stop words in amgsds paths
}
