% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{switch_system}
\alias{switch_system}
\title{Switch AMGSDS system between Oracle and classic ones}
\usage{
switch_system(mode = c("oracle", "classic"), server = NULL)
}
\arguments{
\item{mode}{chr. AMGSDS system mode, "oracle" or "classic" (for commercial users)}

\item{url}{chr. Specific opendap server (for commercial users)}
}
\description{
Switch AMGSDS system between Oracle and classic ones
}
