% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{unfold_array}
\alias{unfold_array}
\title{Conver array to long-format tibble}
\usage{
unfold_array(amgsds_array, time_index = NULL, na.rm = TRUE)
}
\arguments{
\item{amgsds_array}{AMGSDS array object (output = "array")}

\item{time_index}{indices of time to be extracted}

\item{na.rm}{logical. missing values including NaN are removed if \code{TRUE}}
}
\description{
Conver array to long-format tibble
}
