% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_identity_cloud.R
\name{verify_access_token}
\alias{verify_access_token}
\title{Verify validity of access token}
\usage{
verify_access_token(tokens, warn_hour = 6, dtime = 30)
}
\arguments{
\item{tokens}{Oracle tokens}

\item{warn_hour}{warn upcoming re-authorization}

\item{dtime}{small buffer time to suppress errors due to time lag}
}
\description{
Verify validity of access token
}
