% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrmesh.R
\docType{package}
\name{agrmesh-package}
\alias{agrmesh}
\alias{agrmesh-package}
\title{agrmesh: R interface for the Agro-Meteorological Grid Square Data, NARO}
\description{
agrmesh provides R CUI and GUI tools for spatial meteorolgoical data analyses
}
\author{
\strong{Maintainer}: Keach Murakami \email{keach.murakami@gmail.com} (\href{https://orcid.org/0000-0001-8150-9535}{ORCID})

Other contributors:
\itemize{
  \item Yuichi Nagasaki [contributor]
}

}
