% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{amgsds_config}
\alias{amgsds_config}
\title{Verify and initialize id and pw for AMGSDS system}
\usage{
amgsds_config(
  userid = Sys.getenv("amgsds_id"),
  password = Sys.getenv("amgsds_pw"),
  server = "amd.rd.naro.go.jp",
  .force = FALSE
)
}
\arguments{
\item{userid}{your user id issued by the development group}

\item{password}{your password issued by the development group}

\item{server}{URL for OPeNDAP server}

\item{.force}{logical. pass verification for offline .onLoad}
}
\description{
Verify and initialize id and pw for AMGSDS system
}
