% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{replace_stop_words}
\alias{replace_stop_words}
\title{Replace stop words in amgsds paths for WindowsOS}
\usage{
replace_stop_words(str, rev = FALSE)
}
\arguments{
\item{str}{One or more character vectors.}

\item{rev}{Restore replaced words if \code{TRUE}}
}
\description{
Replace stop words in amgsds paths for WindowsOS
}
